Enable CB_FORMAT_PRINTF on LOG(), fix warnings
As LOG() takes a printf-style format string, annotate it with the
CB_FORMAT_PRINTF macro which enables compiler warnings on mismatches
between the format string and arguments. Fix all the usage errors in
ep-engine flagged by this.

Reviewed-on: http://review.couchbase.org/53874
Tested-by: buildbot <build@couchbase.com>

+++ src/executorthread.cc
@@ -111,9 +111,9 @@ void ExecutorThread::run() {
             rel_time_t startReltime = ep_current_time();
             try {
                 LOG(EXTENSION_LOG_DEBUG,
-                    "%s: Run task \"%s\" id %d",
-                getName().c_str(), currentTask->getDescription().c_str(),
-                currentTask->getId());
+                    "%s: Run task \"%s\" id %" PRIu64,
+                    getName().c_str(), currentTask->getDescription().c_str(),
+                    uint64_t(currentTask->getId()));
 
                 // Now Run the Task ....
                 currentTask->setState(TASK_RUNNING, TASK_SNOOZED);
@@ -157,12 +157,12 @@ void ExecutorThread::run() {
                         waketime = new_waketime;
                     }
                     LOG(EXTENSION_LOG_DEBUG, "%s: Reschedule a task"
-                            " \"%s\" id %d[%llu %llu |%llu]",
+                            " \"%s\" id %" PRIu64 "[%" PRIu64 " %" PRIu64 " |%" PRIu64 "]",
                             name.c_str(),
                             currentTask->getDescription().c_str(),
-                            currentTask->getId(), new_waketime,
-                            currentTask->waketime,
-                            waketime);
+                            uint64_t(currentTask->getId()), uint64_t(new_waketime),
+                            uint64_t(currentTask->waketime),
+                            uint64_t(waketime));
                 }
             } catch (std::exception& e) {
                 LOG(EXTENSION_LOG_WARNING,
